backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.0
sqlalchemy==2.0.29
sqlalchemy[asyncio]==2.0.29
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.0
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
anthropic==0.28.0
pdfplumber==0.11.0
pytesseract==0.3.10
pdf2image==1.17.0
openpyxl==3.1.2
python-docx==1.1.0
boto3==1.34.29
minio==7.2.5
celery==5.4.0
redis==5.0.1
pgvector==0.2.0
structlog==24.1.0
slowapi==0.1.8
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
factory-boy==3.3.0
sentry-sdk==1.45.0
